Add update detail page
