--- src/components/Modal/UpdateDetailModal.css
@@ -1,33 +1,0 @@
-.UpdateDetailModal__Title.content {
-  margin-bottom: 36px !important;
-}
-
-.UpdateDetailModal__Content.content {
-  width: 100% !important;
-  padding: 0 52px !important;
-  margin-top: 0 !important;
-}
-
-.UpdateDetailModal__Content .ui.header {
-  font-size: 21px;
-}
-
-.UpdateDetailModal__Content .dg.Paragraph {
-  font-size: 15px;
-  line-height: 24px;
-  overflow-wrap: break-word;
-}
-
-.dg.Paragraph.UpdateDetailModal__Timestamp {
-  color: #736e7d;
-  margin-right: 5px;
-}
-
-.UpdateDetailModal__Date {
-  display: flex;
-  flex-direction: row;
-}
-
-.UpdateDetailModal__Date .dcl.address {
-  font-size: 15px;
-}
