+++ src/components/Section/Divider.css
@@ -1,4 +1,9 @@
 .Divider {
+  width: 100%;
   border: none;
   border-top: 1px solid rgba(115, 110, 125, 0.32);
+}
+
+.Divider__Container--small {
+  width: 40px;
 }
