+++ src/hooks/useProposalUpdate.ts
@@ -1,0 +1,15 @@
+import useAsyncMemo from "decentraland-gatsby/dist/hooks/useAsyncMemo"
+import { Governance } from "../api/Governance"
+
+export default function useProposalUpdate(updateId?: string | null) {
+  const [update, state] = useAsyncMemo(() => Governance.get().getProposalUpdate(updateId!), [updateId], {
+    callWithTruthyDeps: true,
+  })
+
+  console.log(update)
+
+  return {
+    update,
+    state
+  }
+}
